Move utils and tests under utils.rs
