+++ crates/light-client-prover/src/lib.rs
@@ -3,3 +3,4 @@ pub mod circuit;
 pub mod da_block_handler;
 #[cfg(feature = "native")]
 pub mod runner;
+pub(crate) mod utils;
